requirements.txt
easygui==0.98.3
numpy==1.24.2
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
pushbullet.py==0.12.0
skimage==0.0
